[ui] Extract FibsemCanvasBase from FibsemImageCanvas (FIB-409)
Split the canvas into a content-agnostic base and a single-image subclass, so a
second subclass can hold many images placed in stage space (FIB-408) without
changing behaviour under FibsemImageCanvas's ~30 existing consumers.

FibsemCanvasBase owns everything independent of what is drawn: pan/zoom, the
overlay registry and its input gating, the toolbar, the chrome (hint / title /
info bar / LIVE badge / flash / legend), scalebar, crosshair, contrast+gamma and
view fitting. FibsemImageCanvas keeps set_image / set_array / update_display.

The seam is one hook. _fit_view previously read the image artist's extent
directly; it now asks _content_extent(), which the base answers with None and the
image canvas answers with the image's extent. That method is the entire
difference between a canvas holding one image and one holding many.

A pure move otherwise: 54 of 55 methods are byte-identical, module constants and
all four signals are unchanged, and an instance of the new class exposes the same
runtime attribute surface and the same xlim/ylim/extent as the old one for the
same input. No existing test is modified.
